Fix #673
